Fixes for adding, removing, 'current' event
addProduct, removeProduct:
Removed if/else and duplicated code when determining product/timeframe type.

removeProduct:
Divided into removeTimeframe and removeSeries similar to adding.
Implemented declared but missing support for multiple product/timeframe values.
Correct check for the existence of a series.

addTimeframe:
Added check for series existence. When adding series once, this is not necessary, but with repeated calls to addProduct, a situation could arise when the series were overwritten, but the number of event listeners increased.

SetSeriesPrice:
Added declared but missing passing candles to 'current' events.

Also a couple of edits in the readme.